+++ src/main/java/by/koronatech/officeStaffMgmt/api/dto/EmployeeResponseDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @Builder
@@ -12,7 +14,7 @@ import lombok.Setter;
 public class EmployeeResponseDTO implements FullNameProvider {
     private Long id;
     private String fullName;
-    private Float salary;
+    private BigDecimal salary;
     private String department;
     private Boolean manager;
 }
